refactor #7 - get explicity import util methods from their specific modules
