+++ judgments/views/results.py
@@ -1,14 +1,12 @@
-from caselawclient.Client import (
-    MarklogicAPIError
-
-)
+from caselawclient.Client import MarklogicAPIError
 from django.http import Http404, HttpResponse
 from django.template import loader
 from django.utils.translation import gettext
 
 from judgments.models import SearchResult
+from judgments.utils.paginator import paginator
+from judgments.utils.perform_advanced_search import perform_advanced_search
 
-from judgments.utils import perform_advanced_search, paginator
 
 def results(request):
     context = {"page_title": gettext("results.search.title")}
